feat(sam): add support for parameter overrides to execute
