+++ packages/sam/src/builders/cloudformation/deploy/IParameterOverrides.ts
@@ -1,0 +1,3 @@
+export interface IParameterOverrides {
+    [key: string]: string;
+}
