+++ packages/sam/src/builders/execute/options.ts
@@ -1,4 +1,6 @@
 import { JsonObject } from '@angular-devkit/core';
+import { ImportStackOutputs } from '@nx-aws/core';
+import { IParameterOverrides } from '../cloudformation/deploy/IParameterOverrides';
 
 export interface SamExecuteBuilderOptions extends JsonObject {
     args: string[];
@@ -8,4 +10,6 @@ export interface SamExecuteBuilderOptions extends JsonObject {
     host: string;
     port: number;
     mimicEnv: string;
+    importStackOutputs: (ImportStackOutputs & JsonObject) | null;
+    parameterOverrides: IParameterOverrides | null;
 }
