+++ packages/sam/src/utils/loadEnvFromStack.ts
@@ -1,6 +1,7 @@
 import { Observable, of, from } from 'rxjs';
 import { formatStackName } from '@nx-aws/core';
 import { loadEnvironmentVariablesForStackLambdas } from './loadEnvironmentVariablesForStackLambdas';
+
 export function loadEnvFromStack(
     mimicEnv: string | undefined,
     project: string | undefined
